+++ app/src/androidTest/java/nextstep/signup/SignUpScreenTest.kt
@@ -3,7 +3,7 @@ package nextstep.signup
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.domain.SignUpInfo
+import nextstep.signup.ui.SignUpInfo
 import nextstep.signup.ui.SignUpScreen
 import org.junit.Before
 import org.junit.Rule
